Add files via upload
## Description

This Python script, `main_script.py`, is part of the Chicago Data Analysis Project. The project focuses on analyzing Chicago census, public schools, and crime data. The script performs the following tasks:

1. **Data Loading:** Reads data from CSV files (`ChicagoCensusData.csv`, `ChicagoPublicSchools.csv`, `ChicagoCrimeData.csv`) into Pandas DataFrames.

2. **Database Setup:** Creates an SQLite database (`FinalDB.db`) and tables (`table1`, `table2`, `table3`) based on the loaded data.

3. **SQL Queries:** Executes a series of SQL queries on the database to extract meaningful insights. These queries cover various aspects such as demographic analysis, crime statistics, and school-related information.

4. **Results:** Prints the results of the SQL queries to the console, providing valuable information derived from the analyzed datasets.

The script serves as the main entry point for the data analysis process and showcases how Python, Pandas, and SQLite can be utilized for exploring and extracting insights from diverse datasets related to a specific geographic area.